Tách DTO Response & Request, cập nhật category khi thêm mới và cập nhật sản phẩm

+++ src/main/java/com/backend/VNPT_Intern_Project/repositories/ProductRepository.java
@@ -1,6 +1,7 @@
 package com.backend.VNPT_Intern_Project.repositories;
 
-import com.backend.VNPT_Intern_Project.dtos.ProductDTO.ProductDTO;
+import com.backend.VNPT_Intern_Project.dtos.ProductDTO.ProductDTORequest;
+import com.backend.VNPT_Intern_Project.dtos.ProductDTO.ProductDTORsponse;
 import com.backend.VNPT_Intern_Project.repositories.DTOMapper.ProductDTOMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.EmptyResultDataAccessException;
@@ -19,34 +20,37 @@ public class ProductRepository {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    public List<ProductDTO> getAllProducts() {
+    public List<ProductDTORsponse> getAllProducts() {
         String sql = "SELECT * FROM product p";
         return jdbcTemplate.query(sql, new ProductDTOMapper(jdbcTemplate));
     }
 
-    public List<ProductDTO> getProductById(String uuid_product) {
+    public List<ProductDTORsponse> getProductById(String uuid_product) {
         String sql = "SELECT * FROM product p WHERE p.uuid_product = ?";
         return jdbcTemplate.query(sql, new Object[]{uuid_product}, new ProductDTOMapper(jdbcTemplate));
     }
 
-    public List<ProductDTO> getProductsByBrandName(String brand_name) {
-        String sql = "SELECT * FROM brand b " +
+    public List<ProductDTORsponse> getProductsByBrandName(String brand_name) {
+        String sql =
+                "SELECT * FROM brand b " +
                 "LEFT JOIN product p " +
                 "ON b.uuid_brand = p.uuid_brand " +
                 "WHERE b.name = ?";
         return jdbcTemplate.query(sql, new Object[]{brand_name}, new ProductDTOMapper(jdbcTemplate));
     }
 
-    public List<ProductDTO> getProductsByCategoryName(String category_name) {
-        String sql = "SELECT * FROM category c " +
+    public List<ProductDTORsponse> getProductsByCategoryName(String category_name) {
+        String sql =
+                "SELECT * FROM category c " +
                 "LEFT JOIN product_category pc ON c.uuid_category = pc.uuid_category " +
                 "LEFT JOIN product p ON pc.uuid_product = p.uuid_product " +
                 "WHERE c.title = ?";
         return jdbcTemplate.query(sql, new Object[]{category_name}, new ProductDTOMapper(jdbcTemplate));
     }
 
-    public List<ProductDTO> getProductsByBrandAndCategory(String brand_name, String category_name) {
-        String sql = "SELECT * FROM category c " +
+    public List<ProductDTORsponse> getProductsByBrandAndCategory(String brand_name, String category_name) {
+        String sql =
+                "SELECT * FROM category c " +
                 "LEFT JOIN product_category pc ON c.uuid_category = pc.uuid_category " +
                 "LEFT JOIN product p ON pc.uuid_product = p.uuid_product " +
                 "JOIN brand b ON p.uuid_brand = b.uuid_brand " +
@@ -54,11 +58,15 @@ public class ProductRepository {
         return jdbcTemplate.query(sql, new Object[]{brand_name, category_name}, new ProductDTOMapper(jdbcTemplate));
     }
 
-    public int createProduct(ProductDTO product) {
+    public List<ProductDTORsponse> createProduct(ProductDTORequest product) {
         String createProductQuery =
                 "INSERT INTO product (uuid_product, title, meta_title, summary, type, price, quantity, created_date, updated_date, published_date, description, uuid_brand) " +
-                        "VALUES (UUID(), ?, ?, ?, ?, ?, ?, NOW(), NOW(), NOW(), ?, ?)";
+                "VALUES (?, ?, ?, ?, ?, ?, ?, NOW(), NOW(), ?, ?, ?)";
 
+        // Tạo id cho product mới
+        String newProductID = UUID.randomUUID().toString();
+
+        // Kiểm tra brand đã có chưa, chưa thì tạo mới
         String findBrandIdQuery = "SELECT uuid_brand FROM brand WHERE name = ?";
         String brandID = null;
 
@@ -66,26 +74,55 @@ public class ProductRepository {
             brandID = jdbcTemplate.queryForObject(findBrandIdQuery, new Object[]{product.getBrand()}, String.class);
         } catch (EmptyResultDataAccessException e) {
             System.out.println("Brand not found with name: " + product.getBrand());
-            String insertBrandQuery = "INSERT INTO brand (uuid_brand, name, created_date, updated_date) VALUES (?, ?, NOW(), NOW())";
+            String insertBrandQuery =
+                    "INSERT INTO brand (uuid_brand, name, created_date) " +
+                    "VALUES (?, ?, NOW())";
 
             // Tạo UUID cho brand
             brandID = UUID.randomUUID().toString();
-
             jdbcTemplate.update(insertBrandQuery, brandID, product.getBrand());
         }
 
-        return jdbcTemplate.update(createProductQuery,
+        // Thêm sản phẩm vào bảng Product
+        jdbcTemplate.update(createProductQuery,
+                newProductID,
                 product.getTitle(),
                 product.getMetaTitle(),
                 product.getSummary(),
                 product.getType(),
                 product.getPrice(),
                 product.getQuantity(),
+                product.getPublishedDate(),
                 product.getDescription(),
                 brandID);
+
+        // Kiểm tra category, chưa có thì thêm 1 category mới với sản phẩm vừa thêm
+        if(product.getCategory() != null) {
+            String findCategoryQuery =
+                    "SELECT uuid_category FROM category " +
+                    "WHERE title = ?";
+            String categoryID = null;
+            try {
+                categoryID = jdbcTemplate.queryForObject(findCategoryQuery, new Object[]{product.getCategory()}, String.class);
+            } catch (EmptyResultDataAccessException e) {
+                System.out.println("Category not found with name: " + product.getCategory());
+
+                String insertCategory =
+                        "INSERT INTO category (uuid_category, title) " +
+                        "VALUES (?, ?)";
+                categoryID = UUID.randomUUID().toString();
+                jdbcTemplate.update(insertCategory, categoryID, product.getCategory());
+            }
+            String ProductCategoryQuery =
+                    "INSERT INTO product_category (uuid_product, uuid_category) " +
+                            "VALUES (?, ?)";
+            jdbcTemplate.update(ProductCategoryQuery, newProductID, categoryID);
+        }
+
+        return getProductById(newProductID);
     }
 
-    public int updateProduct(ProductDTO product, String uuid_product) {
+    public List<ProductDTORsponse> updateProduct(ProductDTORequest product, String uuid_product) {
         String findProductIDQuery = "SELECT uuid_product FROM product WHERE uuid_product = ?";
         String findBrandIdQuery = "SELECT uuid_brand FROM brand WHERE name = ?";
 
@@ -100,7 +137,7 @@ public class ProductRepository {
             productID = jdbcTemplate.queryForObject(findProductIDQuery, new Object[]{uuid_product}, String.class);
         } catch (EmptyResultDataAccessException e) {
             System.out.println("Không tìm thấy sản phẩm!");
-            return 0;
+            return null;
         }
 
         // Kiểm tra brand đã có chưa, nếu chưa thì tạo mới
@@ -109,7 +146,9 @@ public class ProductRepository {
             brandID = jdbcTemplate.queryForObject(findBrandIdQuery, new Object[]{product.getBrand()}, String.class);
         } catch (EmptyResultDataAccessException e) {
             System.out.println("Brand not found with name: " + product.getBrand());
-            String insertBrandQuery = "INSERT INTO brand (uuid_brand, name, created_date, updated_date) VALUES (?, ?, NOW(), NOW())";
+            String insertBrandQuery =
+                    "INSERT INTO brand (uuid_brand, name, created_date, updated_date) " +
+                    "VALUES (?, ?, NOW(), NOW())";
 
             // Tạo UUID cho brand
             brandID = UUID.randomUUID().toString();
@@ -117,8 +156,7 @@ public class ProductRepository {
             jdbcTemplate.update(insertBrandQuery, brandID, product.getBrand());
         }
 
-        if (productID != null) {
-            return jdbcTemplate.update(updateProductQuery,
+        jdbcTemplate.update(updateProductQuery,
                     product.getTitle(),
                     product.getMetaTitle(),
                     product.getSummary(),
@@ -129,21 +167,64 @@ public class ProductRepository {
                     product.getDescription(),
                     brandID,
                     productID);
+
+        // Kiểm tra category, chưa có thì gắn category với sản phẩm
+        if(product.getCategory() != null) {
+            String findCategoryQuery =
+                    "SELECT uuid_category FROM category " +
+                    "WHERE title = ?";
+            String categoryID = null;
+
+            String ProductCategoryQuery =
+                    "INSERT INTO product_category (uuid_product, uuid_category) " +
+                    "VALUES (?, ?)";
+
+            try {
+                categoryID = jdbcTemplate.queryForObject(findCategoryQuery, new Object[]{product.getCategory()}, String.class);
+            } catch (EmptyResultDataAccessException e) {
+                System.out.println("Category not found with name: " + product.getCategory());
+
+                String insertCategory =
+                        "INSERT INTO category (uuid_category, title) " +
+                        "VALUES (?, ?)";
+                categoryID = UUID.randomUUID().toString();
+                jdbcTemplate.update(insertCategory, categoryID, product.getCategory());
+
+
+                jdbcTemplate.update(ProductCategoryQuery, uuid_product, categoryID);
+            }
+
+            // nếu đã có category, kiểm tra xem sản phẩm đã được gắn với category đó chưa, chưa thì thêm vào product_category
+            String checkProductCategory =
+                    "SELECT uuid_product FROM product_category " +
+                    "WHERE uuid_product = ? AND uuid_category = ?";
+
+            String check = jdbcTemplate.queryForObject(checkProductCategory, new Object[]{uuid_product, categoryID}, String.class);
+
+            if (check == null) {
+                jdbcTemplate.update(ProductCategoryQuery, uuid_product, categoryID);
+            }
         }
-        return 0;
+        return getProductById(productID);
     }
 
-    public int deleteProduct(String uuid_product) {
-        // Xóa sản phẩm ở ProductAttribute
-        String deleteProductAtrributeQuery = "DELETE FROM product_attribute WHERE uuid_product = ?";
-        int deletedPA = jdbcTemplate.update(deleteProductAtrributeQuery, uuid_product);
+    public List<ProductDTORsponse> deleteProduct(String uuid_product) {
+        List<ProductDTORsponse> product = getProductById(uuid_product);
 
-        // Xóa sản phẩm ở ProductCategory
-        String deleteProductCategoryQuery = "DELETE FROM product_category WHERE uuid_product = ?";
-        int deletedCategory = jdbcTemplate.update(deleteProductCategoryQuery, uuid_product);
+        if (product != null) {
+            // Xóa sản phẩm ở ProductAttribute
+            String deleteProductAtrributeQuery = "DELETE FROM product_attribute WHERE uuid_product = ?";
+            jdbcTemplate.update(deleteProductAtrributeQuery, uuid_product);
 
-        // Xóa sản phẩm ở bảng Product
-        String sql = "DELETE FROM product WHERE uuid_product = ?";
-        return jdbcTemplate.update(sql, uuid_product) + deletedPA + deletedCategory;
+            // Xóa sản phẩm ở ProductCategory
+            String deleteProductCategoryQuery = "DELETE FROM product_category WHERE uuid_product = ?";
+            jdbcTemplate.update(deleteProductCategoryQuery, uuid_product);
+
+            // Xóa sản phẩm ở bảng Product
+            String sql = "DELETE FROM product WHERE uuid_product = ?";
+            jdbcTemplate.update(sql, uuid_product);
+        }
+        return product;
     }
 }
+
